Using capacitor dependency to connect firebase android app to extract text from image
Upon clicking async function allows user to choose whether to take a photo or pick from the gallery.

It wraps these 2 options with a cancel button below it through an actionsheet.

Once the user chooses it calls the OCR function after checking if base64Image is valid

The function to extract text from OCR is called and we get the results.